+++ common/params.h
@@ -2,6 +2,7 @@
 
 #include <map>
 #include <string>
+#include <vector>
 
 enum ParamKeyType {
   PERSISTENT = 0x02,
@@ -15,10 +16,11 @@ enum ParamKeyType {
 class Params {
 public:
   Params(const std::string &path = {});
+  std::vector<std::string> allKeys() const;  
   bool checkKey(const std::string &key);
   ParamKeyType getKeyType(const std::string &key);
   inline std::string getParamPath(const std::string &key = {}) {
-    return key.empty() ? params_path + "/d" : params_path + "/d/" + key;
+    return params_path + prefix + (key.empty() ? "" : "/" + key);
   }
 
   // Delete a value
@@ -43,7 +45,7 @@ public:
 
 private:
   std::string params_path;
-
+  std::string prefix;
 
 public:  // atom
   inline int getInt( const char *key )