+++ selfdrive/controls/lib/latcontrol.py
@@ -26,7 +26,7 @@ class LatControl(ABC):
     self.sat_count = 0.
 
   def _check_saturation(self, saturated, CS, steer_limited):
-    if saturated and CS.vEgo > 10. and not not steer_limited and not CS.steeringPressed:
+    if saturated and CS.vEgo > 10. and not steer_limited and not CS.steeringPressed:
       self.sat_count += self.sat_count_rate
     else:
       self.sat_count -= self.sat_count_rate